+++ dojo/tools/cyclonedx/json_parser.py
@@ -4,6 +4,7 @@ import logging
 import dateutil
 from cvss import CVSS3
 from dojo.models import Finding
+from dojo.tools.cyclonedx.helpers import Cyclonedxhelper
 
 LOGGER = logging.getLogger(__name__)
 
@@ -87,7 +88,7 @@ class CycloneDXJSONParser(object):
                         or rating.get("method") == "CVSSv31"
                     ):
                         raw_vector = rating["vector"]
-                        cvssv3 = self._get_cvssv3(raw_vector)
+                        cvssv3 = Cyclonedxhelper()._get_cvssv3(raw_vector)
                         severity = rating.get("severity")
                         if cvssv3:
                             finding.cvssv3 = cvssv3.clean_vector()
@@ -177,15 +178,4 @@ class CycloneDXJSONParser(object):
             components[reference]["version"],
         )
     
-    def _get_cvssv3(self, raw_vector):
-        if raw_vector is None or "" == raw_vector:
-            return None
-        if not raw_vector.startswith("CVSS:3"):
-            raw_vector = "CVSS:3.1/" + raw_vector
-        try:
-            return CVSS3(raw_vector)
-        except BaseException:
-            LOGGER.exception(
-                f"error while parsing vector CVSS v3 {raw_vector}"
-            )
-            return None
+
